requirements.txt
anyio==3.4.0
asgiref==3.4.1
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
backcall==0.2.0
binance-connector==1.7.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
click==8.0.3
constantly==15.1.0
contextlib2==21.6.0
contextvars==2.4
cryptography==35.0.0
dataclasses==0.8
decorator==5.1.0
entrypoints==0.3
fastapi==0.70.0
h11==0.12.0
httptools==0.3.0
hyperlink==21.0.0
idna==2.10
immutables==0.16
importlib-metadata==4.8.2
incremental==21.3.0
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==7.0.6
jupyter-core==4.9.1
mysql-connector-python==8.0.27
nest-asyncio==1.5.1
numpy==1.19.5
pandas==1.1.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pkg_resources==0.0.0
prompt-toolkit==3.0.22
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
Pygments==2.10.0
pyOpenSSL==19.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
requests==2.25.1
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tornado==6.1
traitlets==4.3.3
Twisted==21.2.0
txaio==21.2.1
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
uvloop==0.14.0
watchgod==0.7
wcwidth==0.2.5
websockets==9.1
zipp==3.6.0
zope.interface==5.4.0
